+++ packages/components/src/components/modal/modal-styles.ts
@@ -1,26 +1,25 @@
-import {
-  borderRadiusMedium,
-  gridExtendedOffsetBase,
-  spacingFluidLarge,
-  spacingFluidMedium,
-  spacingFluidSmall,
-} from '@porsche-design-system/styles';
+import { gridExtendedOffsetBase } from '@porsche-design-system/styles';
 import {
   addImportantToEachRule,
   colorSchemeStyles,
-  dismissButtonJssStyle,
   hostHiddenStyles,
   preventFoucOfNestedElementsStyles,
 } from '../../styles';
 import {
+  dialogBorderRadius,
   dialogGridJssStyle,
   dialogHostJssStyle,
-  dialogPaddingBlock,
+  dialogPaddingBottom,
+  dialogPaddingInline,
+  dialogPaddingTop,
   getDialogColorJssStyle,
+  getDialogDismissButtonJssStyle,
   getDialogJssStyle,
-  getDialogStickyFooterJssStyle,
   getDialogTransitionJssStyle,
   getScrollerJssStyle,
+  getSlotFooterJssStyle,
+  getSlotJssStyle,
+  getSlotMainJssStyle,
 } from '../../styles/dialog-styles';
 import type { BreakpointCustomizable } from '../../types';
 import { buildResponsiveStyles, getCss } from '../../utils';
@@ -52,7 +51,6 @@ export const getComponentCss = (
   isOpen: boolean,
   backdrop: ModalBackdrop,
   fullscreen: BreakpointCustomizable<boolean>,
-  hasDismissButton: boolean,
   hasHeader: boolean,
   hasFooter: boolean
 ): string => {
@@ -61,9 +59,9 @@ export const getComponentCss = (
       ':host': {
         display: 'contents',
         ...addImportantToEachRule({
-          [`${cssVarRefPaddingTop}`]: dialogPaddingBlock,
-          [`${cssVarRefPaddingBottom}`]: dialogPaddingBlock,
-          [`${cssVarRefPaddingInline}`]: spacingFluidLarge,
+          [`${cssVarRefPaddingTop}`]: dialogPaddingTop,
+          [`${cssVarRefPaddingBottom}`]: dialogPaddingBottom,
+          [`${cssVarRefPaddingInline}`]: dialogPaddingInline,
           ...dialogHostJssStyle,
           ...colorSchemeStyles,
           ...hostHiddenStyles,
@@ -71,14 +69,8 @@ export const getComponentCss = (
       },
       ...preventFoucOfNestedElementsStyles,
       slot: {
-        display: 'block',
-        '&:first-of-type': {
-          gridRowStart: 1,
-        },
-        '&:not([name])': {
-          gridColumn: '2/3',
-          zIndex: 0, // controls layering + creates new stacking context (prevents content within to be above other dialog areas)
-        },
+        ...getSlotJssStyle(),
+        '&:not([name])': getSlotMainJssStyle(),
         ...(hasHeader && {
           '&[name=header]': {
             gridColumn: '2/3',
@@ -86,7 +78,7 @@ export const getComponentCss = (
           },
         }),
         ...(hasFooter && {
-          '&[name=footer]': getDialogStickyFooterJssStyle(),
+          '&[name=footer]': getSlotFooterJssStyle(),
         }),
       },
       dialog: getDialogJssStyle(isOpen, backdrop),
@@ -112,21 +104,10 @@ export const getComponentCss = (
               maxWidth: '1535.5px', // to be in sync with "Porsche Grid" on viewport >= 1920px: `calc(${gridColumnWidthXXL} * 14 + ${gridGap} * 13)`
               placeSelf: 'center',
               margin: `var(${cssVariableSpacingTop},clamp(16px, 10vh, 192px)) ${gridExtendedOffsetBase} var(${cssVariableSpacingBottom},clamp(16px, 10vh, 192px))`, // horizontal margin is needed to ensure modal is placed on "Porsche Grid" when slotted content is wider than the viewport width
-              borderRadius: borderRadiusMedium,
+              borderRadius: dialogBorderRadius,
             }
       ),
     },
-    ...(hasDismissButton && {
-      dismiss: {
-        ...dismissButtonJssStyle,
-        gridArea: '1/3',
-        zIndex: 2, // ensures dismiss button is above sticky footer, header and content
-        position: 'sticky',
-        insetBlockStart: spacingFluidSmall,
-        marginBlockStart: `calc(${spacingFluidMedium} * -1)`,
-        marginInlineEnd: spacingFluidSmall,
-        placeSelf: 'flex-start flex-end',
-      },
-    }),
+    dismiss: getDialogDismissButtonJssStyle(),
   });
 };